+++ CHANGELOG.md
@@ -17,6 +17,9 @@
 - fix(instance): create instance with an IPv6 #788
 - dedicated-inference/deployment logs: works again even without the --tail argument
 
+## Features
+- feat(cli scaffold): allow deprecated flags #789 
+
 ### Breaking changes
 - instance create: merge --ipv6 and --private-instance into --public-ip #788
 - nodepool: replace --ipv6 and --nodepool-ipv6 with --public-ip and --nodepool-public-ip #788